requirements.txt
groq
python-dotenv
requests
pillow
opencv-python-headless
pandas
numpy==1.24.1
spoonacular
google-generativeai
streamlit
plotly